+++ Flow/src/utopia/flow/collection/mutable/MutableTreeLike.scala
@@ -9,7 +9,7 @@ import utopia.flow.collection
   * @author Mikko Hilpinen
   * @since 1.11.2016
   */
-trait TreeLike[A, NodeType <: TreeLike[A, NodeType]] extends collection.template.TreeLike[A, NodeType]
+trait MutableTreeLike[A, Repr <: MutableTreeLike[A, Repr]] extends collection.template.TreeLike[A, Repr]
 {
 	// ABSTRACT ---------------------
 	
@@ -17,7 +17,7 @@ trait TreeLike[A, NodeType <: TreeLike[A, NodeType]] extends collection.template
 	  * Updates the children of this treelike
 	  * @param newChildren The new children
 	  */
-	protected def setChildren(newChildren: Seq[NodeType]): Unit
+	protected def setChildren(newChildren: Seq[Repr]): Unit
 	
 	
 	// OPERATORS    -----------------
@@ -28,7 +28,7 @@ trait TreeLike[A, NodeType <: TreeLike[A, NodeType]] extends collection.template
 	  * @param child The node that is added under this node
 	  * @return Whether the node was successfully added under this node
 	  */
-	def +=(child: NodeType) =
+	def +=(child: Repr) =
 	{
 		// Makes sure the child doesn't already exist in the direct children
 		// And that this node won't end up under a child node
@@ -45,7 +45,7 @@ trait TreeLike[A, NodeType <: TreeLike[A, NodeType]] extends collection.template
 	  * removed
 	  * @param node The node that is removed from under this node
 	  */
-	def -=(node: Tree[A]): Unit =
+	def -=(node: MutableTree[A]): Unit =
 	{
 		removeChild(node)
 		children.foreach { child => child -= node }
@@ -63,18 +63,18 @@ trait TreeLike[A, NodeType <: TreeLike[A, NodeType]] extends collection.template
 	  * Removes the direct children of this node that match the specified function
 	  * @param f A function that returns true for children that should be removed
 	  */
-	def clearChildrenWhere(f: NodeType => Boolean) = setChildren(children.filterNot(f))
+	def clearChildrenWhere(f: Repr => Boolean) = setChildren(children.filterNot(f))
 	
 	/**
 	  * Removes all children (direct and indirect) based on the specified function result
 	  * @param f A function that returns true for children that should be removed
 	  */
-	def clearNestedWhere(f: NodeType => Boolean): Unit =
+	def clearNestedWhere(f: Repr => Boolean): Unit =
 		setChildren(children.filterNot(f).map { c => c.clearNestedWhere(f); c })
 	
 	/**
 	  * Removes a node from the direct children under this node
 	  * @param child The node that is removed from under this node
 	  */
-	def removeChild(child: TreeLike[A, _]) = setChildren(children.filterNot { _ == child })
+	def removeChild(child: MutableTreeLike[A, _]) = setChildren(children.filterNot { _ == child })
 }